edited delete test case and endpoint

+++ service/routes.py
@@ -124,17 +124,17 @@ def update_accounts(account_id):
 # DELETE AN ACCOUNT
 ######################################################################
 
-@app.route("/accounts/account_id", methods=["DELETE"])
-def delete_accounts():
-    app.logger.info("Request to delete an account")
-    check_content_type("application/json")
+@app.route("/accounts/<int:account_id>", methods=["DELETE"])
+def delete_accounts(account_id):
+    """
+    Delete an Account
+    This endpoint will delete an Account based on the account_id that is requested
+    """
+    app.logger.info("Request to delete an Account with id: %s", account_id)
     account = Account.find(account_id)
-    account.delete()
-    acc = Account.find(account_id)
-    location_url = "/"  # Remove once get_accounts has been implemented
-    return make_response(
-            jsonify(acc), status.HTTP_204_NO_CONTENT, {"Location": location_url}
-        )
+    if account:
+        account.delete()
+    return "", status.HTTP_204_NO_CONTENT
